backend/requirements.txt
fastapi
uvicorn
pandas
matplotlib
mcp
python-multipart
google-generativeai
python-dotenv
seaborn
